requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.1
gradio_client==0.2.7
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.17.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
magic-filter==1.0.11
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
orjson==3.9.7
packaging==23.1
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.7.1
ujson==5.8.0
urllib3==2.0.5
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
